gmail: move send-side provider I/O from client.go into send.go
Send, FindByMessageID, and postJSON pushed client.go to 519 lines, past
make check-backend's hard 500-LOC go-file-length gate. Move the three
functions this task added into send.go instead of waiving the cap:
client.go stays the capture-side Gmail I/O, send.go becomes the whole
outbound path. No behaviour change.
